chore: update testdata, move rpc to custom_reth
