+++ testdata/layer1/account1.keystore/0x0B89A9229940747EAc5918027d3c76968631F304
@@ -1,0 +1,1 @@
+{"crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"d246c0b6aa409dbec802d40a7dda4e20"},"ciphertext":"7086f3f17e254079cca6924e559b1bf58abf52cd504a0c2a2bc31093f348ef1e","kdf":"scrypt","kdfparams":{"dklen":32,"n":8192,"p":1,"r":8,"salt":"c6e5881f372ec129acd5e9ae9fabe560a114e80689d704fd874eb794def55082"},"mac":"116dfc402c4c1284705b68b445e573cb8db5bbbc07161f974aabc9e7aad912bb"},"id":"992073b6-f016-4abb-8398-45c81ad2d8b0","version":3}
