+++ testdata/layer1/account2.keystore/0xF8259389fF5c18034bDa9C136269D79Bd8ab2BD6
@@ -1,0 +1,1 @@
+{"crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"0183cc2d1f7b4582b30c1570ad1a8f55"},"ciphertext":"42e616d01595aa4364f287d0e455802c79e06ffe368d57f7a22f5b58c4090044","kdf":"scrypt","kdfparams":{"dklen":32,"n":8192,"p":1,"r":8,"salt":"713e56751fd89327d4dfeee5297ceaf51f9ca7d59af71d9e560775a466d248db"},"mac":"c0dde26babd342ebca9d7e29ceaae08074904e4417a9475f2c236f9b557df5cc"},"id":"0c097b2c-059f-46f0-b2cf-cbcde8231cae","version":3}
