+++ testdata/layer1/chain.json
@@ -1,0 +1,37 @@
+{
+	"nonce": "0x42",
+	"timestamp": "0x0",
+	"extraData": "0x5343",
+	"gasLimit": "0x1388",
+	"difficulty": "0x400000000",
+	"mixHash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+	"coinbase": "0x0000000000000000000000000000000000000000",
+	"alloc": {
+		"0x0B89A9229940747EAc5918027d3c76968631F304": {
+			"balance": "0x4a47e3c12448f4ad000000"
+		},
+		"0xF8259389fF5c18034bDa9C136269D79Bd8ab2BD6": {
+			"balance": "0x4a47e3c12448f4ad000000"
+		}
+	},
+	"number": "0x0",
+	"gasUsed": "0x0",
+	"parentHash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+	"config": {
+		"ethash": {},
+		"chainId": 12345,
+		"homesteadBlock": 0,
+		"eip150Block": 0,
+		"eip155Block": 0,
+		"eip158Block": 0,
+		"byzantiumBlock": 0,
+		"constantinopleBlock": 0,
+		"petersburgBlock": 0,
+		"istanbulBlock": 0,
+		"berlinBlock": 0,
+		"londonBlock": 0,
+		"terminalTotalDifficulty": 0,
+		"terminalTotalDifficultyPassed": true,
+		"shanghaiTime": 0
+	}
+}
